add protection

+++ bot/iiwb/client/default/admins.py
@@ -98,33 +98,34 @@ class Admins(commands.Cog):
 		member = await ctx.guild.fetch_member(int(_userid))
 		_returnstr = []
 		_count = 0
-		for channel in ctx.guild.channels:
-			if(isinstance(channel, discord.TextChannel)):
-				async for message in channel.history(limit=int(limit)):
-					if message.author.id == _userid:
-						try:
-							newEntry = {
-										"author": message.author.id,
-										"authorname": message.author.name,
-										"channel": message.channel.id,
-										"guild": message.guild.id,
-										"type": str(message.type),
-										"message": message.content,
-										"messageid": message.id,
-										"created": time.mktime(message.created_at.timetuple()),
-										"edited": message.edited_at,
-										"deleted": time.time(),
-										"deletor": ctx.author.id,
-										"deletorname": ctx.author.name
-									}
-							pop = await self.b.insertClearRecord(newEntry)
-						except Exception as e:
-							print(e)
-						_count +=1
-						await message.delete()
-						_returnstr += f"\nMessage supprimé dans {channel}, contre {member}, limit: {limit}"
-		
-		await ctx.send(_returnstr, ephemeral=True)
+		if(await utils.specifiedRole("Cleaner", guild, author, ctx=ctx)):
+			for channel in ctx.guild.channels:
+				if(isinstance(channel, discord.TextChannel)):
+					async for message in channel.history(limit=int(limit)):
+						if message.author.id == _userid:
+							try:
+								newEntry = {
+											"author": message.author.id,
+											"authorname": message.author.name,
+											"channel": message.channel.id,
+											"guild": message.guild.id,
+											"type": str(message.type),
+											"message": message.content,
+											"messageid": message.id,
+											"created": time.mktime(message.created_at.timetuple()),
+											"edited": message.edited_at,
+											"deleted": time.time(),
+											"deletor": ctx.author.id,
+											"deletorname": ctx.author.name
+										}
+								pop = await self.b.insertClearRecord(newEntry)
+							except Exception as e:
+								print(e)
+							_count +=1
+							await message.delete()
+							_returnstr += f"\nMessage supprimé dans {channel}, contre {member}, limit: {limit}"
+			
+			await ctx.send(_returnstr, ephemeral=True)
 
 async def setup(bot):
 	await bot.add_cog(Admins(bot))
